I will not apologize for art.
* Add `lot bpf snippet explain` command

* Restructure the bpfsnippet.go file

* Refactor and fix some of the code
